Move namespace `behavior` into new header `data_types/behavior.hpp`
