+++ simulation/behavior_tree_plugin/include/behavior_tree_plugin/action_node.hpp
@@ -21,7 +21,7 @@
 #include <geometry/spline/catmull_rom_spline.hpp>
 #include <memory>
 #include <string>
-#include <traffic_simulator/data_type/data_types.hpp>
+#include <traffic_simulator/data_type/behavior.hpp>
 #include <traffic_simulator/entity/entity_base.hpp>
 #include <traffic_simulator/hdmap_utils/hdmap_utils.hpp>
 #include <traffic_simulator/helper/stop_watch.hpp>
