+++ simulation/traffic_simulator/include/traffic_simulator/behavior/behavior_plugin_base.hpp
@@ -17,7 +17,7 @@
 
 #include <boost/optional.hpp>
 #include <string>
-#include <traffic_simulator/data_type/data_types.hpp>
+#include <traffic_simulator/data_type/behavior.hpp>
 #include <traffic_simulator/hdmap_utils/hdmap_utils.hpp>
 #include <traffic_simulator/traffic_lights/traffic_light_manager.hpp>
 #include <traffic_simulator_msgs/msg/driver_model.hpp>
