+++ simulation/traffic_simulator/include/traffic_simulator/data_type/data_types.hpp
@@ -23,18 +23,6 @@
 
 namespace traffic_simulator
 {
-namespace behavior
-{
-enum class Request {
-  NONE,
-  LANE_CHANGE,
-  FOLLOW_LANE,
-  WALK_STRAIGHT,
-};
-std::string getRequestString(const Request & value);
-std::ostream & operator<<(std::ostream & stream, const Request & value);
-}  // namespace behavior
-
 namespace speed_change
 {
 enum class Transition {
