+++ simulation/traffic_simulator/src/data_type/data_types.cpp
@@ -16,43 +16,6 @@
 
 namespace traffic_simulator
 {
-namespace behavior
-{
-std::ostream & operator<<(std::ostream & stream, const Request & value)
-{
-  switch (value) {
-    case Request::NONE:
-      stream << "request : NONE" << std::endl;
-      break;
-    case Request::LANE_CHANGE:
-      stream << "request : LANE_CHANGE" << std::endl;
-      break;
-    case Request::FOLLOW_LANE:
-      stream << "request : FOLLOW_LANE" << std::endl;
-      break;
-    case Request::WALK_STRAIGHT:
-      stream << "request : WALK_STRAIGHT" << std::endl;
-      break;
-  }
-  return stream;
-}
-
-std::string getRequestString(const Request & request)
-{
-  switch (request) {
-    case Request::NONE:
-      return "none";
-    case Request::LANE_CHANGE:
-      return "lane_change";
-    case Request::FOLLOW_LANE:
-      return "follow_lane";
-    case Request::WALK_STRAIGHT:
-      return "walk_straight";
-    default:
-      THROW_SEMANTIC_ERROR(request, " is invalid");
-  }
-}
-}  // namespace behavior
 
 namespace speed_change
 {
